requirements.txt
PyNaCl==1.5
discord.py[voice]==2.0
python-dotenv==0.20.0
requests==2.28.0
Pillow==9.2.0
PyAutoGUI==0.9.53
flask==2.0.1
pytest==7.1.3
PyDirectInput==1.0.4
python-libxdo==0.1.2a1
discord-pretty-help==1.3.4 
# python-uinput==0.11.2
